+++ package.json
@@ -65,6 +65,10 @@
     "@storybook/node-logger": "^6.1.21",
     "@storybook/preset-create-react-app": "^3.1.6",
     "@storybook/react": "^6.1.21",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.10",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
     "@testing-library/dom": "^8.11.2",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
@@ -74,6 +78,7 @@
     "lint-staged": "^10.5.3",
     "prettier": "1.19.1",
     "storybook-react-router": "^1.0.8",
+    "typescript": "^4.5.5",
     "webpack": "^4.44.2",
     "webpack-cli": "^4.5.0"
   },
